add _dbm and _gdbm to align naming of error types with runtime
